Remove `clone` method from `Database` trait
Previously, a database acted like a `DatabaseHandle`: a *lightweight*
object you can use to read/write the database, which could be cloned.

Now a database is a fully owned thing, which is not `: Clone` and
requires `&` access for read/write operations. This allows the database
to only deal with synchronization, and leaves the choice of sharing
primitive (`Arc`, or `Rc`, or a scoped `&`) to the client.