Include milliseconds in log timestamps (#688)
There is no readymade method for that so replace `to_rfc2822()` with an
explicit format string which includes milliseconds.

In order to test these changes we add regexes (only for cfg(test))
because the timestamps are formatted according to local timezone which
can differ on machines, so we can't really hardcode any expected values.