Extract block delays into varibles [ECR-1608] (#800)
Initially it was suggested to move these computations into a function,
but actually these are constant values which do not depend on sandbox
state, so we can move these complex values into variables.

I have reviewed uses of sandbox.round_timeout() and PROPOSE_TIMEOUT in
the code, but they seem to be so diverse that it does not make much
sense to give every one a constant. It makes sense for these ones though
for the sake of readability.