Warn about unusually small max_message_length
We have typical default length set to 1 MB for a reason. It is even
mentioned in documentation. Drop a warning if the user (accidentally)
sets this value too low.

I pondered whether we need to output byte sizes in more user-friendly
manner (i.e., as "1 MB" instead of "1048576"), but that's too much of a
hassle for an event which should never happen.