Hard check for max_message_length
Exonum binary serialization format has certain fixed-size components
like message header before the body and a signature after the body.
Using zero-sized bodies is not practical, but it will certainly not fly
if the maximum message size does not allow even that.

Strictly speaking, Exonum will fail to work if the message length does
not allow consesus messages to get through, but accurately accounting
for that size is somewhat a bother. So we give a conservative estimate
of 1024 bytes for the message body.

There are places in code which subtract EMPTY_RESPONSE_SIZE constant
from max_message_len:

    const EMPTY_RESPONSE_SIZE: u32 =
        (HEADER_LENGTH + SIGNATURE_LENGTH +
         2 * PUBLIC_KEY_LENGTH + 8) as u32;

    let unoccupied_message_size =
        self.state.config().consensus.max_message_len -
        EMPTY_RESPONSE_SIZE;

So we certainly should not allow max_message_len to get lower than that
or unsigned integer wraparound will happen. This limit seems to be based
on the size of exonum::messages::protocol::BlockRequest message.