Compute root hashes in-place
Avoid allocating intermediate Vecs by computing the hashes in-place.
The logic is a bit convoluted as we have to account for proper indexing,
but this should be more efficient.

We still do a full slice copy to get the initial Vec, and we may need to
allocate an additional empty hash if the optimizer is not that smart.