Revert "Compute root hashes in-place"
This reverts commit 3e60b21e6a7d79ec1af9c400d79ccc1e501ea358.

After all, I did not really like the optimized in-place implementation.
It has too much magic and weird C-style indexing in it which is not
idiomatic for Rust. I'd guess we'll keep the original implementation
until profiling will show that this is really a bottleneck and we need
to do something about it.