Avoid unnecessary initial Vec allocation
Strictly speaking, we can avoid cloning the whole original Vec if we use
combine_hash_list() to allocate the first Vec with half of the original
hashes.

However, this also requires us to explicitly handle the special case of
hash list of length 1 (we need to simply return the hash instead of
rehashing it).