Inject Proposes on transaction pressure [ECR-1441]
Previously we have tweaked the timeout between Propose messages to
adapt to current size of transaction queue. However, this timeout
is fixed at the start of the round or at block creation time. If
a node suddenly receives sufficiently large number of transactions
during the timeout it will still wait for max_propose_timeout().

Now we are going to check the transaction count every time we add
a new transaction to the queue. If we have enough queued up then
we schedule a quicker Propose timeout with min_propose_timeout().

Previously scheduled timeouts will be handled later but they will be
skipped due the checks in the beginning of handle_propose_timeout().

Also note that we prevent expedited timeouts from being scheduled
repeatedly. This is required as the transactions are not going anyway
out of the queue only because we schedule a timeout. We reset the
flag when we actually handle the nearest timeout.