Add tests for expedited Proposal injection
Here is main course of this patch set. We add some tests to verify
that Proposals are actually injected with minimal timeout if the
transaction pool goes over the threshold. For this we will need to
tweak the default sandbox configuration so that proposal values
make more sense. However, we must take care to not touch the rest
of tests which rely on the default behavior.