Drop tokio-timer dependency [ECR-2120]
Dependabot has been pestering us with update of tokio-timer from 0.1.2
to 0.2.5. However, the API of this crate has changed significantly since
it has become a part of the main Tokio framework. My attempts at using
just tokio-timer alone were not successful. It seems to work only when
tightly integrated with tokio and tokio-core, so let's just use it via
tokio::util::FutureExt::deadline() to set the deadlines we need. With
that we can replace tokio-timer with tokio (which includes the timer).

Tokio is future-based framework so we'll need a Core to run the futures.
Furthermore, futures and streams are not *that* well integrated so in
order to extract an item from a stream with a deadline set we need to
do some dancing around with putting a stream into a future and then
getting it back. Unfortunately, with current API it is not possible to
recover the stream when it times out. But we're panicking in that case
anyway so that should be fine for tests.