Do not reconnect to untrusted validator peers [ECR-2155]
Return value of State::remove_peer_with_addr() is used by
NodeHandler::remove_peer_with_addr() to choose whether to try
reconnecting to the remote verifier peer after the node has been
disconnected from it for whatever reason.

Currently this includes disconnections caused by the fact that our node
does not trust this particular verifier (when the verifier is present
in the shared configuration but is missing from node's local peer list).

In this case we'll still try reconnecting but we will most likely fail
again and produce yet another useless bunch of log lines. The local
configuration will not change without node restart. The remote peer
using an invalid public key is not a transient network error.

To avoid doing needless work, check if the validator is present in our
list of trusted peers before attempting reconnection.

Note that the remote peer is also occasinally trying to connect to us.
so if this was indeed some man-in-the-middle attempt then we'll still
connect to the real peer once the attack is finished.