Do not use words 'untrusted validator'
This terminology has been found confusing to we avoid using it.
The validators are trusted but some of them may be missing from
node's connect list for various reasons. The node does not talk
to nodes that are not on its connect list.

Refactor the interface of remove_peer_with_addr() so that this
behavior is more obvious. Returning a single boolean value is
not adequate for the new behavior. Reconnection decision is
made by NodeHandler and it is based on the information from
State, which now includes not only the presence of a node's
public key in the validator list but the connect list as well.
Add methods to explicitly check the presence after removal.