Do not reconnect to validators not in connect list [ECR-2155] (#894)
* Do not reconnect to untrusted validator peers [ECR-2155]

Return value of State::remove_peer_with_addr() is used by
NodeHandler::remove_peer_with_addr() to choose whether to try
reconnecting to the remote verifier peer after the node has been
disconnected from it for whatever reason.

Currently this includes disconnections caused by the fact that our node
does not trust this particular verifier (when the verifier is present
in the shared configuration but is missing from node's local peer list).

In this case we'll still try reconnecting but we will most likely fail
again and produce yet another useless bunch of log lines. The local
configuration will not change without node restart. The remote peer
using an invalid public key is not a transient network error.

To avoid doing needless work, check if the validator is present in our
list of trusted peers before attempting reconnection.

Note that the remote peer is also occasinally trying to connect to us.
so if this was indeed some man-in-the-middle attempt then we'll still
connect to the real peer once the attack is finished.

* Do not use words 'untrusted validator'

This terminology has been found confusing to we avoid using it.
The validators are trusted but some of them may be missing from
node's connect list for various reasons. The node does not talk
to nodes that are not on its connect list.

Refactor the interface of remove_peer_with_addr() so that this
behavior is more obvious. Returning a single boolean value is
not adequate for the new behavior. Reconnection decision is
made by NodeHandler and it is based on the information from
State, which now includes not only the presence of a node's
public key in the validator list but the connect list as well.
Add methods to explicitly check the presence after removal.