Update actix-web to 0.7.5
A straightforward update, but not without regressions. This was
not mentioned in the changelog or migration guide, but it turned
out that HttpServer::start() now returns Server, not HttpServer.
Addr<HttpServer> is no longer a valid type, we should use the
correct one instead. Remove the type alias altogether as it's
used in only one place and it's not that hard to spell out.