Update actix-web to 0.7.5, remove workarounds [ECR-2149] (#924)
* Update actix-web to 0.7.5

A straightforward update, but not without regressions. This was
not mentioned in the changelog or migration guide, but it turned
out that HttpServer::start() now returns Server, not HttpServer.
Addr<HttpServer> is no longer a valid type, we should use the
correct one instead. Remove the type alias altogether as it's
used in only one place and it's not that hard to spell out.

* Remove scoped path workaround for actix-web 0.7.3 [ECR-2149]

The regression was fixed in 0.7.5 which we currently use. Drop the
workaround which is no longer necessary.