Clean docs before checking for dead links
Travis may cache target directories which may contain leftover
files from previous builds. This may cause false positives in
cargo deadlinks check if an entity is moved or removed, but
not all links to it are updated. If the directory contains the
linked file from some previous build then the link will not be
considered dead while it certainly is.

Regenerating docs from a clean state ensures that we check the
actual version and will not forget to update links.