Fix Server import split off into actix-net
Since v0.7.9 [1] the HTTP server for actix-web is provided by
the actix-net crate. This was not noted in the changelog (as
usual, he-he), but actix_web::server::Server is now available
as actix_net::server::Server.

Add actix-net as our dependency, use the same version 0.1.1
as used by actix-web 0.7.12 at the moment, and import the
Server from that crate.

[1]: https://github.com/actix/actix-web/commit/1907102685a7a1b09a4689b304038f69b8f4b7ef