Skip execution for disabled services
We start introducing the concept of disabled services at its bottom
level: in transaction execution routine that handles every block.
Disabled services are present in the network, have some storage for
them, and by all means are known to nodes. However, they are unable
to execute any transactions or event handlers while disabled.

Therefore, skip before_commit and after_commit handlers for such
services and mark any submitted transactions for them as failed
without even attempting to execute them. This leaves all other
transactions in the block intact while blocking execution for the
disabled services.

We're going to store service status as a global configuration which
must be the same for all nodes. This configuration is not available
before the genesis block is formed. We allow all services to execute
their initialization handlers regardless of their initial enabled
state. For this purpose treat all services as enabled when handling
the genesis block.

This is fine now because the genesis block cannot have any service
transactions in it. However, this might change in the future.