Store service status in global configuration
The global configuration (struct StoredConfiguration) has a dictionary
which keeps service configurations. Extend that entry to store the
service enabled/disabled status.

In order to be compatible with current behavior, treat any service
as enabled by default unless it has been explicitly disabled in the
genesis block configuration.

The genesis block configuration is currently not handled, we're going
to add this in the next commit.