Handle initial service state in genesis configuration
GenesisConfig struct keeps the genesis configuration. (It is stored in
TOML configuration files which nodes read during startup.) Introduce
a new "services" table there to allow the network administrators to
configure the initial enabled status for services.

GenesisConfig will have three optional construction parameters with
this change. It is kinda hard to maintain so replace the constructors
with a builder. As a side effect we may now easily see all places in
the code where we deal with genesis configuration and update them all.

As with StoredConfiguration, make the configuration optional and treat
services as enabled unless they are explicitly disabled by the admins.

It may look like a lax policy but it is also quite convenient. Note
how many places use genesis configuration. If we were to enforce
whitelisting of enabled services instead of blacklisting of disabled
ones then all these users will need to know the exhaustive list of
services in the network. It would be much harder to thread the service
list to all these places.