Populate initial service status with command extension
Unfortunately, the command-line handling code has no direct access to
node instances and does not know about the list of available services.
Thus it is not possible to generate the default initial service state
in GenerateCommonConfig command.

We get around this restriction by installing a custom command extension
that keeps the list of names of all available services and generates
a suitable default value for the table.